[Refactor] state 과부하를 줄였다. (use 커스텀 훅 추가)
